Move getAllTablenamesWithoutFiltering to CassandraDao
Also, Made cassandraDao a field in CassandraKVS.
